requirements.txt
aiogram==3.4.1
anthropic==0.21.3
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.3
apscheduler==3.10.4
Pillow==10.2.0
fpdf2==2.7.9
pytz==2024.1
